Add repository and change primary key to long
Przy użyciu zmiennej typu long przy implementacji primary key zwiększamy ilość dostępnych miejsc .

+++ PAI-EVENT-REST/src/main/java/pl/RK/PAIEVENTREST/models/Comment.java
@@ -7,7 +7,7 @@ import java.util.Date;
 public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int commentId;
+    long commentId;
 
     String text;
